Armature: Only normalize with first bone weight
OK, so this is a little confusing. For some reason, PangYa mpet files store vertices 'denormalized' as I'm calling it- in this context, I'm using it to mean it's stored such that multiplying it against the bone matrix gives the neutral position. It's not really clear why this is done, but one might assume it's to simplify the game code. Except that can't be it.

This change underscores why: it turns out only the _first_ bone weight matters for this denormalization. That means even the game itself must first multiply the vertex by the relevent bone matrix, because vertices with multiple weights will not work properly.

I can't imagine why this is. It may just be cruft from earlier code that did not support bone weights.